Remove ElasticClientSetup
