+++ search/src/main/scala/weco/api/search/Main.scala
@@ -3,15 +3,17 @@ package weco.api.search
 import org.apache.pekko.http.scaladsl.server.Directives.concat
 import org.apache.pekko.actor.ActorSystem
 import com.typesafe.config.Config
+import weco.Tracing
 import weco.api.search.config.MultiClusterConfigParser
 import weco.api.search.config.builders.PipelineElasticClientBuilder
 import weco.api.search.elasticsearch.ResilientElasticClient
-import weco.api.search.models.ApiConfig
+import weco.api.search.models.{ApiConfig, ApiEnvironment, ElasticConfig, PipelineClusterElasticConfig}
 import weco.typesafe.WellcomeTypesafeApp
 import weco.http.WellcomeHttpApp
 import weco.http.monitoring.HttpMetrics
 import weco.http.typesafe.HTTPServerBuilder
 import weco.monitoring.typesafe.CloudWatchBuilder
+import weco.typesafe.config.builders.EnrichConfig.RichConfig
 
 object Main extends WellcomeTypesafeApp {
 
@@ -21,12 +23,40 @@ object Main extends WellcomeTypesafeApp {
     implicit val actorSystem: ActorSystem = ActorSystem("search-api")
     implicit val ec: scala.concurrent.ExecutionContext = actorSystem.dispatcher
 
-    val (elasticClient, elasticConfig) =
-      ElasticClientSetup.buildDefaultElasticClientAndConfig(
-        config = config,
-        serviceName = "catalogue_api",
-        environment = apiConfig.environment
-      )
+    val (elasticClient, elasticConfig) = apiConfig.environment match {
+      case ApiEnvironment.Dev =>
+        info(s"Running in dev mode.")
+        val pipelineDateOverride = config.getStringOption("dev.pipelineDate")
+        val pipelineDate =
+          pipelineDateOverride.getOrElse(ElasticConfig.pipelineDate)
+        if (pipelineDateOverride.isDefined)
+          warn(s"Overridden pipeline date: $pipelineDate")
+        (
+          new ResilientElasticClient(
+            clientFactory = () =>
+              PipelineElasticClientBuilder(
+                serviceName = "catalogue_api",
+                pipelineDate = pipelineDate,
+                environment = apiConfig.environment
+              )),
+          PipelineClusterElasticConfig(
+            config.getStringOption("dev.pipelineDate")
+          )
+        )
+      case _ =>
+        info(s"Running in deployed mode (environment=$apiConfig.environment)")
+        // Only initialise tracing in deployed environments
+        Tracing.init(config)
+        (
+          new ResilientElasticClient(
+            clientFactory = () =>
+              PipelineElasticClientBuilder(
+                serviceName = "catalogue_api",
+                environment = apiConfig.environment
+              )),
+          PipelineClusterElasticConfig()
+        )
+    }
 
     // Parse multi-cluster configuration
     val additionalClusterConfigs =
